Add configurable S3 logging
We were unable to get enough logging out of S3 using such a Bauhaus
implementation in the storage service's S3 adapter. The Boto3 logging
will explode this so that there is a workable level of logging that
can be used to debug issues with different S3 implementations in the
future. This logging should only ever be used with test data.